+++ CHANGELOG.md
@@ -19,8 +19,10 @@ All notable changes to this project will be documented in this file.
 - Package updates
 
 ### 5.2.4
-- Adding emojis to languages and platforms
+- Added structure for using emojis for languages and platforms
 - Fixed improper handling of non-existent media in `m-anime`, `m-manga`, `m-animesearch`, and `m-mangasearch`
+- `m-booru` compatibility fix
+- Minor background changes
 - Package updates
 
 ## 5.1 - Are Visual Novels Games? (2020-04-22)